move run button functions into separate file
